Refactoring axios-retry into a plugin
Functions could not be used from the nuxt/module module options.  Also, the retry would not be persisted when new clients were created, like the Craft client.
