+++ modules/axios.js
@@ -3,7 +3,7 @@
  */
 import defaultsDeep from 'lodash/defaultsDeep'
 import { requireLate } from '@cloak-app/utils'
-import { isNetworkError, isRetryableError } from 'axios-retry'
+import { join } from 'path'
 export default function() {
 
 	// Set default config
@@ -15,36 +15,14 @@ export default function() {
 		// Don't trigger loader UI
 		progress: false,
 
-		// Retry failed network requests up to 3 times
-		retry: { retries: 3, retryCondition },
 	}})
+
+	// Add custom axios-retry functionality
+	this.addPlugin({
+		src: join(__dirname, '../plugins/axios-retry.js')
+	})
 
 	// Add the Axios Nuxt module late so it can be used in plugins
 	requireLate(this, '@nuxtjs/axios')
 }
 
-// Retry GQL requests that don't have a mutation
-function retryCondition(error) {
-	return isNetworkError(error) ||
-		isIdempotentRequestError(error) ||
-		isGQLQuery(error)
-}
-
-// Allow POST requests that look like GQL queries (aka, non-mutations)
-function isGQLQuery(error) {
-
-	// Only act on POSTs
-	if (error.config.method != 'post') return
-
-	// Test whether the request body is non-empty
-	const data = error.config?.data
-	if (!data) return
-
-	// ... and contains JSON
-	let query
-	try { query = JSON.parse(data).query }
-	catch (e) { return }
-
-	// Ensure the request is not a mutation and that the error is retryable
-	return !query.includes('mutation') && isRetryableError(error)
-}
